refactor(cli): extract TypeChecker to separate module
